[local-sync] Stop sync worker before deleting account database
Summary:
Various errors are thrown when the sync worker tries accessing
a database that we've already deleted, so make sure the sync
worker has been stopped before we remove the database. This diff
involves modifying `Interruptible` so that `interrupt()` returns
a promise that resolves once the interrupt has been completed.

Addresses T7472

Test Plan: manual

Reviewers: evan, juan

Reviewed By: evan, juan

Differential Revision: https://phab.nylas.com/D3679

+++ packages/local-sync/src/local-sync-worker/index.js
@@ -7,6 +7,8 @@ const manager = require('./sync-process-manager')
 // Account objects. We want to sync all K2 Accounts, but when an N1 Account is
 // deleted, we want to delete the K2 account too.
 
+const deletionsInProgress = new Set();
+
 async function ensureK2Consistency() {
   const {Account} = await LocalDatabaseConnector.forShared();
   const k2Accounts = await Account.findAll();
@@ -16,11 +18,13 @@ async function ensureK2Consistency() {
   const deletions = [];
   for (const k2Account of k2Accounts) {
     const deleted = !n1Emails.includes(k2Account.emailAddress);
-    if (deleted) {
+    if (deleted && !deletionsInProgress.has(k2Account.id)) {
       console.warn(`Deleting K2 account ID ${k2Account.id} which could not be matched to an N1 account.`)
-      manager.removeWorkerForAccountId(k2Account.id);
+      deletionsInProgress.add(k2Account.id)
+      await manager.removeWorkerForAccountId(k2Account.id);
       LocalDatabaseConnector.destroyAccountDatabase(k2Account.id);
-      deletions.push(k2Account.destroy());
+      const deletion = k2Account.destroy().then(() => deletionsInProgress.delete(k2Account.id))
+      deletions.push(deletion)
     }
   }
   return await Promise.all(deletions)
